refactor: move dolphin version response definition to version check
