+++ src/dolphin/types.ts
@@ -38,12 +38,3 @@ export interface PlayKey {
   displayName: string;
   latestVersion?: string;
 }
-
-export interface DolphinVersionResponse {
-  version: string;
-  downloadUrls: {
-    darwin: string;
-    linux: string;
-    win32: string;
-  };
-}
